+++ js/test/units/testunitfmt_mt_MT.js
@@ -1,7 +1,7 @@
 /*
  * testunitfmt_mt_MT.js - test the unitfmt for mt-MT
  *
- * Copyright © 2021,2023 JEDLSoft
+ * Copyright © 2021, 2023, 2026 JEDLSoft
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -88,7 +88,27 @@ module.exports.testunitfmt_mt_MT = {
             length: "long"
         });
         var str = uf.format(m1);
-        test.equal(str, "-16.666666666666668 grad Celsius");
+        var platform = ilib._getPlatform();
+        if (platform === "nodejs") {
+            var cldrVersion = Number(process.versions["cldr"]);
+            if (cldrVersion < 36) {
+                // no Intl.PluralRules support here, so iLib's own plural rules
+                // are used; the absolute value of -16.6 selects the "many"
+                // category (n mod 100 = 16, in 11..19) -> "-il grad"
+                test.equal(str, "-16.666666666666668-il grad Celsius");
+            } else {
+                test.equal(str, "-16.666666666666668 grad Celsius");
+            }
+        } else if (platform === "browser") {
+            var cldrVersion = TestingSupport.getCLDRVersionForBrowser();
+            if (cldrVersion !== undefined && cldrVersion >= 36) {
+                test.equal(str, "-16.666666666666668 grad Celsius");
+            } else {
+                test.equal(str, "-16.666666666666668-il grad Celsius");
+            }
+        } else {
+            test.equal(str, "-16.666666666666668-il grad Celsius");
+        }
         test.done();
     },
     testUnitFormatArea1_mt_MT: function(test) {